+++ src/alpha/u8x4/sse4.rs
@@ -202,19 +202,33 @@ unsafe fn divide_alpha_4_pixels(src_pixels: __m128i) -> __m128i {
     let alpha_scale = _mm_set1_ps(255.0 * 256.0);
 
     let alpha_f32 = _mm_cvtepi32_ps(_mm_srli_epi32::<24>(src_pixels));
-    let scaled_alpha_f32 = _mm_div_ps(alpha_scale, alpha_f32);
-    let scaled_alpha_i32 = _mm_cvtps_epi32(scaled_alpha_f32);
-    let mma0 = _mm_shuffle_epi8(scaled_alpha_i32, shuffle1);
-    let mma1 = _mm_shuffle_epi8(scaled_alpha_i32, shuffle2);
+    let recip_alpha_i32 = _mm_cvtps_epi32(_mm_div_ps(alpha_scale, alpha_f32));
+    // Recip alpha in Q8.8 format
+    let recip_alpha_lo_q8_8 = _mm_shuffle_epi8(recip_alpha_i32, shuffle1);
+    let recip_alpha_hi_q8_8 = _mm_shuffle_epi8(recip_alpha_i32, shuffle2);
 
-    let pix0 = _mm_unpacklo_epi8(zero, src_pixels);
-    let pix1 = _mm_unpackhi_epi8(zero, src_pixels);
+    // Pixels components in format Q9.7
+    let components_lo_q9_7 = _mm_slli_epi16::<7>(_mm_unpacklo_epi8(src_pixels, zero));
+    let components_hi_q9_7 = _mm_slli_epi16::<7>(_mm_unpackhi_epi8(src_pixels, zero));
 
-    let pix0 = _mm_mulhi_epu16(pix0, mma0);
-    let pix1 = _mm_mulhi_epu16(pix1, mma1);
+    // Multiplied pixels components as i16.
+    //
+    // fn _mm_mulhrs_epi16(a: i16, b: i16) -> i16 {
+    //   let tmp: i32 = ((a as i32 * b as i32) >> 14) + 1;
+    //   (tmp >> 1) as i16
+    // }
+    let max_value = _mm_set1_epi16(0xff);
+    let res_components_lo_i16 = _mm_min_epu16(
+        _mm_mulhrs_epi16(components_lo_q9_7, recip_alpha_lo_q8_8),
+        max_value,
+    );
+    let res_components_hi_i16 = _mm_min_epu16(
+        _mm_mulhrs_epi16(components_hi_q9_7, recip_alpha_hi_q8_8),
+        max_value,
+    );
 
     let alpha = _mm_and_si128(src_pixels, alpha_mask);
-    let rgb = _mm_packus_epi16(pix0, pix1);
+    let rgba = _mm_packus_epi16(res_components_lo_i16, res_components_hi_i16);
 
-    _mm_blendv_epi8(rgb, alpha, alpha_mask)
+    _mm_blendv_epi8(rgba, alpha, alpha_mask)
 }